fix(hr.payslip): Set partner_id on all account.move lines for consistet behavior with account.invoic
When creating the account.move.line objects resulted by validating a payslip, the partner_id field

where just being filled when the salary.rule have a corresponding register partner or on receivable

and payable accounts. This is a different bahavior than the one we are using on the account.invoice

where we set the partner on all lines. This different behavior confuse users thus we are changing

the behavior to better consistency